+++ src/main/java/com/test/demo/repository/CartRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends CrudRepository <CartItem, Integer> {
+public interface CartRepository extends CrudRepository <CartItem, Long> {
 
     
 }
